update About page
