+++ client/src/pages/AboutOwner.js
@@ -1,0 +1,29 @@
+import React from 'react';
+
+function AboutOwner() {
+    return (
+        <>
+            <p>
+                You need a break. So do your dogs. We will help both! 
+                Woof connects dog owners with dog walkers who’ll treat your dog like family.
+                You can trust us to keep your pet happy, healthy, and sweet as ever.
+            </p>
+            <p>
+                We understand that life can be hectic and it may be hard to fit in a lengthy walk for your furry friend. 
+                We are here to take that stress away from you to give your pal the exercise and 
+                relief that it needs so you can focus on other tasks at hand.
+            </p>
+            <p>
+                We are dog lovers and our dog walkers have lots of experience taking care of dogs. 
+                Walkers are responsible and caring. They take care of your pet with respect. 
+                They offer affection and attention to your pet, when you can’t be there.
+            </p>
+            <p>
+                Woof is also a cutting edge technology business committed to making pet care safe, 
+                easy, and affordable so that everyone can experience the unconditional love of a pet.
+            </p>
+        </>
+    )
+}
+
+export default AboutOwner;
