+++ client/src/pages/AboutWalker.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function AboutWalker() {
+    return (
+        <>
+            About Walker
+        </>
+    )
+}
+
+export default AboutWalker;
